requirements.txt
streamlit
huggingface_hub
openai
chromadb
tiktoken
langchain
gradio
pyngrok
pypdf
InstructorEmbedding
sentence_transformers